refactor: make log fixture re-usable across packages
